# add AES-SIV-CMAC amd64 assembler implementation