--- a/.travis.yml
+++ b/.travis.yml
@@ -19,8 +19,10 @@ branches:
 before_script:
 - go mod download
 - diff -au <(go mod verify) <(printf "all modules verified\n")
+- go get -u github.com/klauspost/asmfmt/cmd/asmfmt
 
 script:
 - diff -au <(gofmt -d .) <(printf "")
+- diff -au <(asmfmt -d .) <(printf "")
 - diff -au <(go vet -all .) <(printf "")
 - go test -v ./...