--- a/aes_cmac_amd64.go
+++ b/aes_cmac_amd64.go
@@ -7,11 +7,61 @@
 package siv
 
 import (
+	"crypto/subtle"
+	"hash"
+
+	cmac "github.com/aead/cmac/aes"
 	"golang.org/x/sys/cpu"
 )
 
-func newCMAC(key []byte) authEnc {
+func keySchedule(keys, key *byte, keyLen uint64)
+
+func xorKeyStream(dst, src, iv, keys []byte, keyLen uint64)
+
+func newCMAC(key []byte) aead {
 	if cpu.X86.HasAES {
+		cmac, _ := cmac.New(key[:len(key)/2])
+		key = key[len(key)/2:]
+		keys := make([]byte, 4*(28+len(key)))
+		keySchedule(&keys[0], &key[0], uint64(len(key)))
+		return &aesSivCMacAsm{
+			cmac:      cmac,
+			keys:      keys,
+			keyLength: len(key),
+		}
 	}
 	return newCMACGeneric(key)
+}
+
+type aesSivCMacAsm struct {
+	cmac      hash.Hash
+	keys      []byte
+	keyLength int
+}
+
+func (c *aesSivCMacAsm) seal(ciphertext, nonce, plaintext, additionalData []byte) {
+	v := s2vGeneric(additionalData, nonce, plaintext, c.cmac)
+	copy(ciphertext, v[:])
+	ciphertext = ciphertext[len(v):]
+
+	iv := newIV(v)
+	xorKeyStream(ciphertext, plaintext, iv[:], c.keys, uint64(c.keyLength))
+}
+
+func (c *aesSivCMacAsm) open(plaintext, nonce, ciphertext, additionalData []byte) error {
+	var v [16]byte
+	copy(v[:], ciphertext)
+	ciphertext = ciphertext[len(v):]
+
+	iv := newIV(v)
+	xorKeyStream(plaintext, ciphertext, iv[:], c.keys, uint64(c.keyLength))
+
+	tag := s2vGeneric(additionalData, nonce, plaintext, c.cmac)
+	if subtle.ConstantTimeCompare(v[:], tag[:]) != 1 {
+		for i := range plaintext {
+			plaintext[i] = 0
+		}
+		return errOpen
+	}
+	return nil
 }